Transport in ifs is now in [kg/s], and uu is now a real variable. /Jocke
git-svn-id: http://svn.tracmass.org/trunk@229 2ee5d717-d111-0410-bb1d-d2df4af4126a

git-svn-id: https://github.com/TRACMASS/tracmass.git@190 dd566b57-2d5a-cdae-10d7-ce57b7db7bff

+++ trunk/src/loop_pos.f95
@@ -21,7 +21,7 @@ contains
   subroutine  pos(ia,iam,ja,ka,ib,jb,kb,x0,y0,z0,x1,y1,z1)
     
     INTEGER                                    :: mra,mta,msa
-    INTEGER                                    :: uu
+    REAL                                       :: uu
     INTEGER                                    :: ia, iam, ja, ka
     INTEGER                                    :: ib, jb, kb
     REAL                                       :: temp,salt,dens
